+++ src/main/java/com/appdynamics/extensions/solr/mbeans/MBeansHandler.java
@@ -12,6 +12,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -29,26 +30,26 @@ public class MBeansHandler {
         this.httpClient = httpClient;
     }
 
-    public Map<String, Long> populateStats(Core core, String contextRoot, String serverUrl) {
-        Map<String,Long> mBeansMetrics = Maps.newHashMap();
+    public Map<String, BigDecimal> populateStats(Core core, String contextRoot, String serverUrl) {
+        Map<String, BigDecimal> mBeansMetrics = Maps.newHashMap();
         String url = buildUrl(core, contextRoot, serverUrl);
         Map<String, JsonNode> solrStatsMap = buildSolrMBeansMap(core, url);
         //core metrics
         mBeansMetrics.putAll(new CoreMetrics(core.getName()).populateStats(solrStatsMap));
         //query metrics
         for (String handler : core.getQueryHandlers()) {
-            mBeansMetrics.putAll(new QueryMetrics(core.getName()).populateStats(solrStatsMap,handler));
+            mBeansMetrics.putAll(new QueryMetrics(core.getName()).populateStats(solrStatsMap, handler));
         }
         //cache metrics
         mBeansMetrics.putAll(new CacheMetrics(core.getName()).populateStats(solrStatsMap));
         return mBeansMetrics;
     }
 
-    private Map<String,JsonNode> buildSolrMBeansMap(Core core,String url){
+    private Map<String, JsonNode> buildSolrMBeansMap(Core core, String url) {
         CloseableHttpResponse response = null;
         Map<String, JsonNode> solrStatsMap = null;
         try {
-            logger.debug("fetching solr mbean handler map from {}",url);
+            logger.debug("fetching solr mbean handler map from {}", url);
             response = HttpHelper.doGet(httpClient, url);
             solrStatsMap = parseResponse(core, response);
         } catch (Exception e) {
@@ -63,12 +64,13 @@ public class MBeansHandler {
      * The different metrics for cache, core, query are found in the json object following the name of the mBean handler.
      * For eg.
      * [
-     *   "Core",
-     *   {},
-     *   "Cache",
-     *   {}
+     * "Core",
+     * {},
+     * "Cache",
+     * {}
      * ]
      * This method parses the response.
+     *
      * @param core
      * @param response
      * @return
@@ -89,7 +91,7 @@ public class MBeansHandler {
         return solrStatsMap;
     }
 
-    private String buildUrl(Core core, String contextRoot, String serverUrl){
+    private String buildUrl(Core core, String contextRoot, String serverUrl) {
         StringBuilder url = new StringBuilder(serverUrl);
         url.append(contextRoot).append(String.format(MBEANS_PATH, core.getName()));
         return url.toString();
