+++ src/main/java/com/appdynamics/extensions/solr/mbeans/cache/CacheMetrics.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -15,14 +16,14 @@ public class CacheMetrics {
         this.coreName = coreName;
     }
 
-    public Map<String, Long> populateStats(Map<String, JsonNode> solrMBeansHandlersMap) {
-        Map<String, Long> cacheMetrics = new HashMap<String, Long>();
+    public Map<String, BigDecimal> populateStats(Map<String, JsonNode> solrMBeansHandlersMap) {
+        Map<String, BigDecimal> cacheMetrics = new HashMap<String, BigDecimal>();
         try {
             cacheMetrics.putAll(new DocumentCacheMetrics(coreName).populateStats(solrMBeansHandlersMap));
             cacheMetrics.putAll(new QueryCacheMetrics(coreName).populateStats(solrMBeansHandlersMap));
             cacheMetrics.putAll(new FieldCacheMetrics(coreName).populateStats(solrMBeansHandlersMap));
             cacheMetrics.putAll(new FilterCacheMetrics(coreName).populateStats(solrMBeansHandlersMap));
-        }catch (Exception e) {
+        } catch (Exception e) {
             logger.error("Error Retrieving cache Stats for " + coreName, e);
         }
         return cacheMetrics;
